requirements.txt
pydantic>=2.5.0
pyyaml>=6.0.1
typer>=0.9.0
openpyxl>=3.1.2
pyvis>=0.3.2
pytest>=7.4.0
pytest-cov>=4.1.0
